feat(Server): header verification

+++ server/include/Server/InternalClient.hpp
@@ -33,9 +33,9 @@ class InternalClient
         void shouldDisconnect(bool _disconnect);
         [[nodiscard]] bool shouldDisconnect() const;
 
-        void setSeqNumber(size_t _seqnum);
+        void setSeqNumber(uint32_t _seqnum);
         void nextSeqNumber();
-        [[nodiscard]] size_t getSeqNumber() const;
+        [[nodiscard]] uint32_t getSeqNumber() const;
 
         [[nodiscard]] std::shared_ptr<net::INetTcp> getSocket() const;
 
@@ -49,7 +49,7 @@ class InternalClient
         bool m_logged_in = false;
         bool m_should_dc = false;
         UserId m_user_id = "";
-        size_t m_seq_num = 0;
+        uint32_t m_seq_num = 0;
 
         std::shared_ptr<net::INetTcp> m_socket = nullptr;
         std::unordered_map<size_t, std::chrono::system_clock::time_point> m_request{};
